a few results including: a longitudinal simulation, nuclear decay process simulation, and data analysis of beam study experiment
long_dynamics_simulation: a simple longitudinal dynamics simulaton and
plots 1) the phi-W phasespace according to a Hamiltonian 2) and the
longitudinal tracking of a few particles in the phi-W phasespace --
visualizing synchrotron oscillation.

side_project_kiwamu: a fun little comparison of FORTRAN Euler method
code vs Python RK4 code on a system of ODEs. The situation is a parent
isotope is being bombarded by neutrons and splits into daughter
isotopes. The application is to use this process to take radioactive
isotopes with a very long half - life to daughter isotopes that are more
stable / less radioactive.

width_delay_beam_experiment:
This is an analysis of a beam studies experiment, folder contains a
report and result plots.